+++ apps/io-func-sign-issuer/src/app/use-cases/__test__/close-signature-request.spec.ts
@@ -164,7 +164,9 @@ const closeSignatureRequest = (input: ClosedSignatureRequest) =>
     telemetryService,
     notificationService,
     eventAnalyticsClient: eventProducerClient,
+    legacyBillingEventProducer: eventProducerClient,
     billingEventProducer: eventProducerClient,
+    legacyEventAnalyticsClient: eventProducerClient,
     inputDecoder: ClosedSignatureRequest,
     logger: {
       log: () => () => void 0,
@@ -212,7 +214,7 @@ describe("closeSignatureRequest", () => {
         mocks.signatureRequest.signed
       )();
       expect(E.isRight(result));
-      expect(analytics.sendBatch).toHaveBeenCalledTimes(2);
+      expect(analytics.sendBatch).toHaveBeenCalledTimes(4);
     });
 
     it("fails on error on billing event", async () => {