+++ apps/io-func-sign-issuer/src/app/use-cases/validate-upload.ts
@@ -210,7 +210,6 @@ export const validateUpload = (
           RTE.chainW(({ documentMetadata }) =>
             pipe(
               getUploadedDocumentUrl(meta.id),
-              RTE.fromReader,
               RTE.chainW(createDocumentFromUrl(meta.documentId)),
               RTE.chainEitherKW((url) =>
                 pipe(