+++ apps/io-func-sign-issuer/src/infra/azure/functions/set-signature-request-status.ts
@@ -23,6 +23,7 @@ type SetSignatureRequestStatusDeps = {
   issuerRepository: IssuerRepository;
   signatureRequestRepository: SignatureRequestRepository;
   eventAnalyticsClient: EventProducerClient;
+  legacyEventAnalyticsClient: EventProducerClient; // WEU — rimuovere post-migrazione
   ready: QueueClient;
   updated: QueueClient;
 };