+++ infra/italynorth/prod/main.tf
@@ -125,3 +125,12 @@ module "storage_io_sign" {
 
   tags = local.tags
 }
+
+module "event_hub" {
+  source = "../_modules/event_hub"
+
+  vnet_common_name_itn           = local.vnet_common_name_itn
+  common_resource_group_name_itn = local.common_resource_group_name_itn
+
+  tags = local.tags
+}