+++ packages/io-sign/src/infra/azure/__test__/event-hubs.spec.ts
@@ -43,6 +43,7 @@ const signatureRequest: SignatureRequestSigned = {
 };
 
 const createAndSendAnalyticsEvent = makeCreateAndSendAnalyticsEvent(
+  undefined as unknown as EventHubProducerClient,
   undefined as unknown as EventHubProducerClient
 );