+++ apps/io-sign-backoffice-app/CHANGELOG.md
@@ -1,5 +1,39 @@
 # io-sign-backoffice-app
 
+## 1.8.5
+
+### Patch Changes
+
+- Updated dependencies [bcbd02c]
+  - @io-sign/io-sign@1.10.0
+
+## 1.8.4
+
+### Patch Changes
+
+- Updated dependencies [7b64d31]
+  - @io-sign/io-sign@1.9.0
+
+## 1.8.3
+
+### Patch Changes
+
+- Updated dependencies [fb57d04]
+  - @io-sign/io-sign@1.8.0
+
+## 1.8.2
+
+### Patch Changes
+
+- Updated dependencies [2593e4c]
+  - @io-sign/io-sign@1.7.0
+
+## 1.8.1
+
+### Patch Changes
+
+- f625285: Remove postbuild command
+
 ## 1.8.0
 
 ### Minor Changes